update crud
status, crud orders cakes(masih ada cacat logic)

+++ src/components/Cakes/List.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from "react";
 import axios from "axios";
-import { Link } from "react-router-dom";
+import { Link, NavLink } from "react-router-dom";
+import Swal from "sweetalert2";
 
 export default function ListCakes() {
     const [cakes, setCakes] = useState([]);
@@ -20,18 +21,34 @@ export default function ListCakes() {
     };
 
     const deleteCake = async (id) => {
-        try {
-            await axios.delete(`https://delivery-cake-api.vercel.app/api/api/cakes/${id}`);
-            setCakes(cakes.filter((cake) => cake.id !== id));
-        } catch (error) {
-            setError("Error: " + error.message);
-        }
+        Swal.fire({
+            title: 'Apakah kamu yakin?',
+            text: "Kue akan dihapus secara permanen!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Ya, hapus!',
+            cancelButtonText: 'Batal'
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                try {
+                    await axios.delete(`https://delivery-cake-api.vercel.app/api/api/cakes/${id}`);
+                    setCakes(cakes.filter((cake) => cake.id !== id));
+                    Swal.fire('Terhapus!', 'Kue telah berhasil dihapus.', 'success');
+                } catch (error) {
+                    setError("Error: " + error.message);
+                }
+            }
+        });
     };
 
     return (
         <div className="container mt-5">
             <h2 className="mb-4">List of Cakes</h2>
             {error && <div className="alert alert-danger">{error}</div>}
+
+            <NavLink to="/Cakes/create" className="btn btn-primary mb-3">Tambah Kue</NavLink>
 
             <table className="table">
                 <thead>
